app/requirements.txt
fastapi
uvicorn
autogen
pydantic
python-dotenv
